+++ lib/ui/screens/school_life/index.dart
