+++ lib/ui/screens/stats/widgets/leading_icon.dart
