+++ lib/ui/screens/summary/index.dart
@@ -17,10 +17,10 @@ import 'package:ynotes/ui/components/y_page/y_page.dart';
 import 'package:ynotes/ui/components/y_page/y_page_local.dart';
 import 'package:ynotes/ui/mixins/layout_mixin.dart';
 import 'package:ynotes/ui/screens/grades/index.dart';
-import 'package:ynotes/ui/screens/summary/summaryPageWidgets/quickGrades.dart';
-import 'package:ynotes/ui/screens/summary/summaryPageWidgets/quickHomework.dart';
-import 'package:ynotes/ui/screens/summary/summaryPageWidgets/quickSchoolLife.dart';
-import 'package:ynotes/ui/screens/summary/summaryPageWidgets/summaryPageSettings.dart';
+import 'widgets/quick_grades.dart';
+import 'widgets/quick_homework.dart';
+import 'widgets/quick_school_life.dart';
+import 'sub_pages/settings.dart';
 
 Future? donePercentFuture;
