+++ lib/ui/screens/summary/widgets/quick_school_life.dart
